requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
jinja2
python-dotenv
markdown
python-multipart 
pyodbc
streamlit

langchain
langchain-google-genai
langgraph
